requirements.txt
python-dotenv>=1.0.0
pydantic>=2.5.0
anthropic>=0.79.0
httpx>=0.27.0
newsapi-python>=0.2.7
feedparser>=6.0.0
python-telegram-bot>=22.5
aiosqlite>=0.20.0
matplotlib>=3.8.0
tenacity>=8.2.0
